[KOGITO-8321] Added bash script to execute a build using Buildah
