+++ builder/kubernetes/testdata/Dockerfile.buildah
@@ -1,0 +1,33 @@
+#Not yet released Change the 31 st oct with quay.io/repository/kiegroup/kogito-swf-builder:latest
+FROM quay.io/kiegroup/kogito-swf-builder-nightly:latest AS builder
+
+# Kogito User
+USER 1001
+
+# User home from base image
+WORKDIR /home/kogito/kogito-sw-base
+
+# Copy from build context to skeleton project
+ADD https://github.com/kiegroup/container-builder/blob/fbf8ef217a694377747749ebfc728149dfb2b471/builder/testdata/greetings.sw.json ./src/main/resources
+
+# Maven vars enhirited from the base image
+RUN /home/kogito/launch/build-app.sh ./resources
+
+#=============================
+# Runtime Run
+#=============================
+FROM registry.access.redhat.com/ubi8/openjdk-11:1.11
+
+ENV LANG='en_US.UTF-8' LANGUAGE='en_US:en'
+
+# We make four distinct layers so if there are application changes the library layers can be re-used
+COPY --from=builder --chown=185 /home/kogito/kogito-sw-base/target/quarkus-app/lib/ /deployments/lib/
+COPY --from=builder --chown=185 /home/kogito/kogito-sw-base/target/quarkus-app/*.jar /deployments/
+COPY --from=builder --chown=185 /home/kogito/kogito-sw-base/target/quarkus-app/app/ /deployments/app/
+COPY --from=builder --chown=185 /home/kogito/kogito-sw-base/target/quarkus-app/quarkus/ /deployments/quarkus/
+
+EXPOSE 8080
+USER 185
+ENV AB_JOLOKIA_OFF=""
+ENV JAVA_OPTS="-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager"
+ENV JAVA_APP_JAR="/deployments/quarkus-run.jar"
